Fixed bug file missing

+++ app/jobs/delete_database.rb
@@ -5,18 +5,21 @@ class DeleteDatabase < RnaAsAPIInteractor
   end
 
   def self.all
-    File.delete(SaveLastMonthlyStockNames.new.path_import)
-    File.delete(SaveLastMonthlyStockNames.new.path_waldec)
+    path_import = SaveLastMonthlyStockNames.new.path_import
+    path_waldec = SaveLastMonthlyStockNames.new.path_waldec
+
+    File.delete(path_import) if File.exist?(path_import)
+    File.delete(path_waldec) if File.exist?(path_waldec)
     Association.delete_all
   end
 
   def self.only_waldec
-    File.delete(SaveLastMonthlyStockNames.new.path_waldec)
+    File.delete(SaveLastMonthlyStockNames.new.path_waldec) if File.exist?(SaveLastMonthlyStockNames.new.path_waldec)
     Association.where(is_waldec: true).delete_all
   end
 
   def self.only_import
-    File.delete(SaveLastMonthlyStockNames.new.path_import)
+    File.delete(SaveLastMonthlyStockNames.new.path_import) if File.exist?(SaveLastMonthlyStockNames.new.path_import)
     Association.where(is_waldec: false).delete_all
   end
 end
